cppcheck warnings fixed

+++ lab8-1/src/input.c
@@ -72,12 +72,12 @@ ReturnCode scan_validate_edge(Edge *edge, int vertices_count)
 
 void graph_add_edge(Graph *graph, Edge *edge)
 {
-  int *graph_array = graph->graph_array;
+  uint *graph_array = graph->graph_array;
   int vertices_count = graph->vertices_count;
   int src = edge->src;
   int dst = edge->dst;
-  graph_array[(src - 1) * vertices_count + (dst - 1)] = edge->weight;
-  graph_array[(dst - 1) * vertices_count + (src - 1)] = edge->weight;
+  graph_array[(src - 1) * vertices_count + (dst - 1)] = (uint)edge->weight;
+  graph_array[(dst - 1) * vertices_count + (src - 1)] = (uint)edge->weight;
 }
 
 ReturnCode scan_validate_edges(Graph *graph)
